% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_crossing_ways.R
\name{find_crossing_ways}
\alias{find_crossing_ways}
\title{find_crossing_ways() Searches for crossing and not connected (high)ways}
\usage{
find_crossing_ways(highways = NULL,
                          exclude = NULL,
                          exclude_bridges = TRUE,
                          return_all_columns = FALSE)
}
\arguments{
\item{highways}{simple feature collection (class sf) with highways as LINESTRING}

\item{exclude}{list of types of highways to exclude, ex. c("proposed", "rest_area", "steps"), default NULL}

\item{exclude_bridges}{if bridges (and highways on any \code{layer}) has to be excluded, default TRUE}

\item{return_all_columns}{if all columns from input data set has to be returned, default FALSE. In this case only osm_id, group, highway, name, and geometry are returned.}
}
\value{
set of highways which crosses but are not connected if any found, otherwise NULL
}
\description{
find_crossing_ways() Searches for crossing and not connected (high)ways
}
\examples{
\dontrun{
ways <- system.file("extdata/highways.gpkg", package = "osmvalidator") |>
  sf::st_read()

find_crossing_ways(highways = ways,
                   exclude = c("proposed", "rest_area", "bus_stop", "steps"))
}

}
