% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unconnected_ways.R
\name{find_unconnected_ways}
\alias{find_unconnected_ways}
\title{find_unconnected_ways() Searches for unconnected solitary ways}
\usage{
find_unconnected_ways(highways = NULL, exclude = NULL, return_all_columns = FALSE)
}
\arguments{
\item{highways}{simple feature collection (class sf) with highways as LINESTRING}

\item{exclude}{list of types of highways to exclude, ex. c("proposed", "rest_area", "steps"), default NULL}

\item{return_all_columns}{if all columns from input data set has to be returned, default FALSE. In this case only osm_id, highway, name, and geometry are returned.}
}
\value{
set of unconnected highways, otherwise NULL
}
\description{
find_unconnected_ways() Searches for unconnected solitary ways
}
\examples{
\dontrun{
ways <- system.file("extdata/highways.gpkg", package = "osmvalidator") |>
  sf::st_read()

find_unconnected_ways(ways)

}

}
